Seance 17

+++ todo/app/Http/Controllers/TaskController.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Category;
 use Dompdf\Dompdf;
+use SimpleSoftwareIO\QrCode\Facades\QrCode;
 
 
 class TaskController extends Controller
@@ -133,6 +134,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('delete-task');
         $task->delete();
 
         return redirect()->route('task.index')->withSuccess('Task Deleted Successfully!');
@@ -145,9 +147,11 @@ class TaskController extends Controller
 
     public function pdf(Task $task)
     {
+        $qrCode = QrCode::size(200)->generate(route('task.show', $task->id));
+        // return $qrCode;
         $pdf = new Dompdf();
         $pdf->setPaper('letter', 'portrait');
-        $pdf->loadHtml(view('task.show-pdf', ["task" => $task]));
+        $pdf->loadHtml(view('task.show-pdf', ["task" => $task, "qrCode" => $qrCode]));
         $pdf->render();
         return $pdf->stream('task_'.$task->id.'.pdf');
     }
